requirements.txt
asgiref==3.7.2
Django==3.2.2
django-ckeditor==6.7.0
django-sendgrid-v5==1.2.2
et-xmlfile==1.1.0
gunicorn==21.2.0
numpy==1.21.6
openpyxl==3.1.2
packaging==23.1
pandas==1.3.5
Pillow==9.5.0
python-decouple==3.8
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2023.3.post1
sendgrid==6.10.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.2.0

